move schemas into module
